refactor to separate code into ADR-specified files/folders
