+++ api.planx.uk/modules/gis/docs.yaml
@@ -2,7 +2,7 @@ openapi: 3.1.0
 info:
   title: Plan✕ API
   version: 0.1.0
-tags: 
+tags:
   - name: gis
 paths:
   /lpa:
